% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_add.R
\name{example_add}
\alias{example_add}
\title{Add two numbers}
\usage{
example_add(a, b)
}
\arguments{
\item{a}{a number}

\item{b}{a number}
}
\value{
a number
}
\description{
Take two numbers as input and output the sum of the two numbers
}
\examples{
example_add(4, 10)
}
